Alpha 2 start
Started on making the website more user friendly. Making it so the client askes the server for the categories in "linje" and them putting them in a select list for piking. I've only started on this and is not complete as of this commit. Have to save it for later.

Made some small changes to the training.Only making the synthetic data set smaller and getting a score of 80%.