Alpha 2.2
Added so "fullt" now gets dynamic list from the server backend category list.

Training now has config variable SAVE_AS_BEST. This is for overwriting the better than the previous best check that it does. This is intended to be used when the model has very new data and has to sort of reset the best model due to big data changes.